requirements.txt
python-pptx
httpx
